sweep: move mocks into one file
